services/api-gateway/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-multipart==0.0.9
redis[hiredis]==5.0.1
nats-py==2.7.2
grpcio==1.62.0
grpcio-tools==1.62.0
protobuf==4.25.3
httpx==0.27.0
pyotp==2.9.0
qrcode==7.4.2
ldap3==2.9.1
python3-saml==1.16.0
cryptography==42.0.4
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-instrumentation-fastapi==0.44b0
prometheus-client==0.20.0
structlog==24.1.0
websockets==12.0
pytest==8.0.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
ruff==0.2.2
